feat: create PUT method to update rental

+++ backend/src/main/java/com/openclassrooms/chatop/model/Rental.java
@@ -1,9 +1,6 @@
 package com.openclassrooms.chatop.model;
 
 import jakarta.persistence.*;
-import lombok.Data;
-import lombok.Getter;
-import lombok.Setter;
 
 import java.time.LocalDate;
 
@@ -12,7 +9,7 @@ import java.time.LocalDate;
 public class Rental {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     private String name;
     private Double surface;
@@ -29,11 +26,11 @@ public class Rental {
     @Column(name = "updated_at")
     private LocalDate updatedAt;
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
